refactor: duplicate teh shared code for better dx
