+++ package.json
@@ -27,13 +27,12 @@
   },
   "homepage": "https://github.com/vuejs/vitepress/tree/master/#readme",
   "scripts": {
-    "dev": "run-p dev-client dev-client-copy dev-node dev-shared",
+    "dev": "run-p dev-client dev-client-copy dev-node",
     "dev-client": "tsc -w -p src/client",
     "dev-client-copy": "node scripts/watchAndCopy",
     "dev-node": "tsc -w -p src/node",
-    "dev-shared": "tsc -w -p src/shared",
     "release": "bash scripts/release.sh",
-    "build": "rimraf -rf dist && tsc -p src/client && tsc -p src/node && tsc -p src/shared && node scripts/copy",
+    "build": "rimraf -rf dist && tsc -p src/client && tsc -p src/node && node scripts/copy",
     "changelog": "conventional-changelog -p angular -i CHANGELOG.md -s"
   },
   "engines": {
