+++ src/node/tsconfig.json
@@ -1,10 +1,14 @@
 {
   "extends": "../../tsconfig.base.json",
   "compilerOptions": {
-    "outDir": "../../dist",
+    "baseUrl": ".",
+    "outDir": "../../dist/node",
     "module": "commonjs",
     "lib": ["ESNext", "DOM"],
     "sourceMap": true
   },
-  "include": [".", "../shared", "../../types/shared.d.ts"]
+  "include": [
+    ".",
+    "../../types/shared.d.ts"
+  ]
 }
