--- src/shared/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "baseUrl": ".",
-    "outDir": "../../dist/client/shared",
-    "module": "esnext",
-    "lib": ["ESNext", "DOM"],
-  },
-  "include": [
-    ".", 
-    "../../types/shared.d.ts",
-  ]
-}
